new source code merged

+++ submodules/SettingsUI/Sources/Data and Storage/DataAndStorageSettingsController.swift
@@ -338,11 +338,11 @@ private enum DataAndStorageEntry: ItemListNodeEntry {
                 return ItemListSectionHeaderItem(presentationData: presentationData, text: text, sectionId: self.section)
             case let .autoplayGifs(_, text, value):
                 //CloudVeil start
-               let v = value && !MainController.SecurityStaticSettings.disableAutoPlayGifs
+                let v = value && !MainController.SecurityStaticSettings.disableAutoPlayGifs
                 return ItemListSwitchItem(presentationData: presentationData, title: text, value: v, sectionId: self.section, style: .blocks, updated: { value in
                     arguments.toggleAutoplayGifs(value)
                 }, tag: DataAndStorageEntryTag.autoplayGifs)
-               //CloudVeil end
+                //CloudVeil end
             case let .autoplayVideos(_, text, value):
                 return ItemListSwitchItem(presentationData: presentationData, title: text, value: value, sectionId: self.section, style: .blocks, updated: { value in
                     arguments.toggleAutoplayVideos(value)
@@ -430,20 +430,20 @@ private func stringForUseLessDataSetting(_ dataSaving: VoiceCallDataSaving, stri
     }
 }
 
-private func stringForAutoDownloadTypes(strings: PresentationStrings, decimalSeparator: String, photo: Bool, videoSize: Int32?, fileSize: Int32?) -> String {
+private func stringForAutoDownloadTypes(strings: PresentationStrings, decimalSeparator: String, photo: Bool, videoSize: Int64?, fileSize: Int64?) -> String {
     var types: [String] = []
     if photo && videoSize == nil {
         types.append(strings.ChatSettings_AutoDownloadSettings_TypePhoto)
     }
     if let videoSize = videoSize {
         if photo {
-            types.append(strings.ChatSettings_AutoDownloadSettings_TypeMedia(autodownloadDataSizeString(Int64(videoSize), decimalSeparator: decimalSeparator)).string)
+            types.append(strings.ChatSettings_AutoDownloadSettings_TypeMedia(autodownloadDataSizeString(videoSize, decimalSeparator: decimalSeparator)).string)
         } else {
-            types.append(strings.ChatSettings_AutoDownloadSettings_TypeVideo(autodownloadDataSizeString(Int64(videoSize), decimalSeparator: decimalSeparator)).string)
+            types.append(strings.ChatSettings_AutoDownloadSettings_TypeVideo(autodownloadDataSizeString(videoSize, decimalSeparator: decimalSeparator)).string)
         }
     }
     if let fileSize = fileSize {
-        types.append(strings.ChatSettings_AutoDownloadSettings_TypeFile(autodownloadDataSizeString(Int64(fileSize), decimalSeparator: decimalSeparator)).string)
+        types.append(strings.ChatSettings_AutoDownloadSettings_TypeFile(autodownloadDataSizeString(fileSize, decimalSeparator: decimalSeparator)).string)
     }
 
     if types.isEmpty {
@@ -641,12 +641,12 @@ public func dataAndStorageController(context: AccountContext, focusOnItemTag: Da
         }).start()
     }, toggleAutoplayGifs: { value in
         let _ = updateMediaDownloadSettingsInteractively(accountManager: context.sharedContext.accountManager, { settings in
-              var settings = settings
-              //CloudVeil start
-              settings.autoplayGifs = value && !MainController.SecurityStaticSettings.disableAutoPlayGifs
-              //CloudVeil end
-              return settings
-          }).start()
+            var settings = settings
+            //CloudVeil start
+            settings.autoplayGifs = value && !MainController.SecurityStaticSettings.disableAutoPlayGifs
+            //CloudVeil end
+            return settings
+        }).start()
     }, toggleAutoplayVideos: { value in
         let _ = updateMediaDownloadSettingsInteractively(accountManager: context.sharedContext.accountManager, { settings in
             var settings = settings
